refactor(plugins): move PluginSettingsCache into a /components folder
