+++ public/app/features/plugins/components/AppRootPage.tsx
@@ -4,7 +4,7 @@ import { AppEvents, AppPlugin, AppPluginMeta, KeyValue, NavModel, PluginType } f
 import { createHtmlPortalNode, InPortal, OutPortal, HtmlPortalNode } from 'react-reverse-portal';
 
 import Page from 'app/core/components/Page/Page';
-import { getPluginSettings } from '../PluginSettingsCache';
+import { getPluginSettings } from './PluginSettingsCache';
 import { importAppPlugin } from '../plugin_loader';
 import { getNotFoundNav, getWarningNav, getExceptionNav } from 'app/angular/services/nav_model_srv';
 import { appEvents } from 'app/core/core';
