+++ public/app/features/plugins/components/PluginSettingsCache.ts
