+++ public/app/features/plugins/tests/AppRootPage.test.tsx
@@ -1,7 +1,7 @@
 import { act, render, screen } from '@testing-library/react';
 import React, { Component } from 'react';
 import AppRootPage from '../components/AppRootPage';
-import { getPluginSettings } from '../PluginSettingsCache';
+import { getPluginSettings } from '../components/PluginSettingsCache';
 import { importAppPlugin } from '../plugin_loader';
 import { getMockPlugin } from '../__mocks__/pluginMocks';
 import { AppPlugin, PluginType, AppRootProps, NavModelItem } from '@grafana/data';
